Added util classes Added MoviesListApi.kt Added KinopoiskAuthInterceptor.kt Added MoviesListInteractor.kt Added required mappers and models Added basic UI Added APiErrorHandling

+++ app/src/main/java/com/pashacabu/testapp/data/model/RatingModel.kt
@@ -1,6 +1,6 @@
 package com.pashacabu.testapp.data.model
 
 data class RatingModel(
-    val imdb: Int? = null,
-    val kp: Int? = null
+    val imdb: Float? = null,
+    val kp: Float? = null
 )
